Find view files using templated names
Sorry, that's a lousy summary. The idea is that, instead of dynamically
querying the Cell class hierarchy for the names of possible templates at
render time, a Cell subclass gathers an array of inherited name
templates. This is a different use of "template", and I'd welcome a
better name. Name templates are just strings like:

'#{attr1}/foo/#{attr2}'

which are interpolated (eval'ed) at render time in the context of the
Cell instance.

This is intended as a powerful replacement for the instance method view
name override, which, if you used it, precluded falling back on the
inherited template file names.

Also in this commit is a marker for abstract Cell classes.

The tests and documentation have not yet been updated, but I think will
be once the implementation has shown itself stable.